A lot of updates
Gone through every section and subsection and fixed the errors made in
the peer-review-version.
Added a lot of images in the Task Design section and updated the rig
design image.
Update Survey
Added intro in Appendix
Change format from list to sub subsections in Task Design due to the
amount of text and images.